+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Remove ability to set mime types with a .types file
 - Add ability to override mime type and charset lookup with globally-set
   functions
+- Removes default charset of utf8 - if you need this, try using a custom
+  charset lookup function
 
 2019/02/10 Version 3.3.1
 - Publish via linux to hopefully fix #238