+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   functions
 - Removes default charset of utf8 - if you need this, try using a custom
   charset lookup function
+- Update modules and fix linting
 
 2019/02/10 Version 3.3.1
 - Publish via linux to hopefully fix #238