+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-- Drop testing/support for nodes 4 and 5
+- Drop testing/support for nodes 4 and 5, test nodes 9, 10 and 11
 - Fix parsing of CORS options
 - Upgrade mime module to v2, use charset module for charset detection
 - Remove ability to set mime types with a .types file